% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/generics.R, R/job.R
\name{browse.jenkinsConnection}
\alias{browse.jenkinsConnection}
\alias{browse}
\alias{browse.character}
\alias{browse.jenkinsJob}
\title{Load resource into an HTML Browser}
\usage{
\method{browse}{jenkinsConnection}(x, ...)

browse(x, ...)

\method{browse}{character}(x, ...)

\method{browse}{jenkinsJob}(x, ...)
}
\arguments{
\item{x}{resource to browse}

\item{...}{further arguments to \code{\link[utils]{browseURL}}}
}
\description{
S3 generic wrapper replacement of \code{\link[utils]{browseURL}}
}
