% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{createJob}
\alias{createJob}
\title{Create a new job}
\usage{
createJob(jenkins, name, config)
}
\arguments{
\item{jenkins}{Jenkins server configuration. Result of \code{\link{jenkins}}}

\item{name}{job name}

\item{config}{job XML specification: \itemize{
\item given as a file path
\item or an object returned by \code{\link{pipelineConfig}} or \code{\link{multibranchPipelineConfig}}
\item or any object of class \code{xml_document} from the \code{xml2} package
}}
}
\value{
jenkins job
}
\description{
Creates a new job with given name and XML config.
}
\seealso{
\link[xml2]{read_xml} \link{crumbRequest}

\link{createMultiBranchPipeline} \link{createPipeline}
}
