% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobConfig.R
\name{multibranchPipelineConfig}
\alias{multibranchPipelineConfig}
\alias{gitBranchSource}
\alias{gitHubBranchSource}
\title{Create a new MultiBranch Pipeline}
\usage{
multibranchPipelineConfig(branchSource)

gitBranchSource(remote, credentialsId)

gitHubBranchSource(
  owner,
  repository,
  credentialsId,
  sshCredentialsId = NULL,
  api = "https://api.github.com"
)
}
\arguments{
\item{credentialsId}{credentials used to scan branches and check out sources}

\item{owner}{GitHub repository owner}

\item{repository}{GitHub repository name}

\item{api}{GitHub api URI}

\item{sshCheckoutCredentialsId}{(optional) if not \code{NULL}, the id of the
ssh credentials to use for the git checkout.}
}
\value{
XML document
}
\description{
Creates a new multibranch pipeline with given name and
git branch source. More info can be found \href{https://wiki.jenkins.io/display/JENKINS/Pipeline+Multibranch+Plugin}{here}.
}
\examples{

if (interactive()) {
  
  # jc <- jenkinsConnection(...)
  
  if (hasJob(jc, "my-job"))
    deleteJob(getJob(jc, "my-job"))
  
  createMultiBranchPipeline(
      jc,
      name = "my-job",
      branchSource = gitBranchSource(
          remote = "https://path/to/my_repository.git",
          credentialsId = "credentials-to-use"))
  
  createMultiBranchPipeline(jc,
      name = "my-job",
      branchSource = gitHubBranchSource(
          api = "https://api.github.com",
          owner = "OpenAnalytics",
          repository = "rjenkins",
          credentialsId = "oajenkins-github"))
  
}
}
\seealso{
\link{createPipeline} \link{createJob}
}
