% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packamon.R
\name{createPackamonJob}
\alias{createPackamonJob}
\title{Create a Packamon job for the given git repository}
\usage{
createPackamonJob(conn, remote = "origin", branch = "master",
  description = "Automatically created. Managed by packamon.")
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \link{jenkins}}

\item{remote}{either a remote name or a repository url. If a remote name
is supplied, the repository url will be looked up from the git config.}

\item{branch}{branch name}

\item{description}{job description}
}
\value{
job name
}
\description{
Create a Packamon job for the given git repository
}
