% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJob.R
\name{createPipeline}
\alias{createPipeline}
\title{Create a new Pipeline}
\usage{
createPipeline(conn, name, remote, credentialsId)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \code{\link{jenkinsConnection}}}

\item{name}{job name}

\item{remote}{the git remote repository to check out}

\item{credentialsId}{credentials used to scan branches and check out sources}
}
\value{
jenkins job
}
\description{
Creates a new pipeline job with given name and git remote
details.
}
