% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{crumbHeader}
\alias{crumbHeader}
\title{Create a crumb header}
\usage{
crumbHeader(crumb)
}
\arguments{
\item{crumb}{result of \link{crumbRequest}}
}
\value{
request header, argument to \code{\link[httr]{POST}}
}
\description{
Create a crumb header
}
