% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{endLines}
\alias{endLines}
\title{Add newlines to the end of text snippets}
\usage{
endLines(texts)
}
\arguments{
\item{texts}{\code{character()} vector of text snippets}
}
\value{
\code{character()}
}
\description{
Add newlines to the end of text snippets
}
