% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{escapeJenkinsItemName}
\alias{escapeJenkinsItemName}
\title{Ensure that names are safe to build URL's with}
\usage{
escapeJenkinsItemName(name)
}
\arguments{
\item{name}{item name}
}
\value{
safe item name
}
\description{
Ensure that names are safe to build URL's with
}
