% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_jenkins.R
\name{extractPackageArchives}
\alias{extractPackageArchives}
\title{Extract Package Archives Names from Build Artifacts}
\usage{
extractPackageArchives(artifacts, latestOnly = TRUE,
  archivePattern = "^([a-zA-Z0-9.]{2,})_([0-9]+\\\\.[0-9]+[.-][0-9]+)\\\\.tar\\\\.gz$")
}
\arguments{
\item{artifacts}{\code{character()} vector of artifact name(s)}

\item{latestOnly}{retain only the latest version of multiple versions of the
same package exist.}

\item{archivePattern}{the regex used for filtering and extracting package
information}
}
\value{
\code{data.frame} with valid package archive names and their
corresponding package name and version number.
}
\description{
Filter a given vector of artifact names to only retain the
desired package archive names.
}
\examples{
artifacts <- c("eightFive_3.4-3.tar.gz", "xgsldgks03", "test_0.3-2.tar.gz",
    "test_1.3-2.tar.gz", "eightFive_3.4-2.tar.gz")
show(extractPackageArchives(artifacts))
##                  archive      name version
## 1 eightFive_3.4-3.tar.gz eightFive   3.4-3
## 3      test_1.3-2.tar.gz      test   1.3-2
}
\seealso{
\link{https://cran.r-project.org/doc/manuals/r-devel/R-exts.html#Creating-R-packages}
}
