% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groovy.R
\name{formatArgument}
\alias{formatArgument}
\title{Format arguments and parameters}
\usage{
formatArgument(value)
}
\arguments{
\item{value}{value to format}
}
\value{
\code{character()}
}
\description{
Format arguments and parameters
}
