% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groovy.R
\name{formatGroovyClosure}
\alias{formatGroovyClosure}
\title{Format a groovy closure}
\usage{
formatGroovyClosure(closure)
}
\arguments{
\item{closure}{\code{\link{GroovyClosure}}}
}
\value{
\code{character()}
}
\description{
Format a groovy closure
}
