% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\name{formatParameter}
\alias{formatParameter}
\title{Format arguments and parameters}
\usage{
formatParameter(value)
}
\arguments{
\item{value}{value to format}
}
\value{
\code{character()}
}
\description{
DSL helper function
}
