% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{getBuildHistory}
\alias{getBuildHistory}
\title{Get the Build History}
\usage{
getBuildHistory(job, depth = 3)
}
\arguments{
\item{job}{object of class \code{jenkinsJob} e.g. created using \code{\link{getJob}}}

\item{depth}{maximum number of builds to list}
}
\value{
\code{data.frame} where each row corresponds to one build
}
\description{
Get the Build History
}
