% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{getBuildLog}
\alias{getBuildLog}
\title{Retrieve the build log}
\usage{
getBuildLog(conn, job, build = "lastBuild", start = 0)
}
\arguments{
\item{conn}{result of \link{jenkinsConnection}}

\item{job}{job name}

\item{build}{build number of identifier}

\item{start}{byte offset}
}
\description{
Retrieve the build log
}
