% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{getJob}
\alias{getJob}
\title{Find a job on jenkins}
\usage{
getJob(conn, name)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \code{\link{jenkinsConnection}}}

\item{name}{job name}
}
\value{
object of class \code{jenkinsJob}
}
\description{
Find a job on jenkins
}
