% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{getJobInfo}
\alias{getJobInfo}
\title{Get General Job Information}
\usage{
getJobInfo(job)
}
\arguments{
\item{job}{object of class \code{jenkinsJob} e.g. created using \code{\link{getJob}}}
}
\value{
list with properties
}
\description{
Query some general information about a job such as it's
descriptione etc.
}
