% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groovy.R
\name{GroovyCall}
\alias{GroovyCall}
\title{Groovy Method Call}
\usage{
GroovyCall(.name, ...)
}
\arguments{
\item{.name}{method name}

\item{...}{mix of unnamed and named arguments: \code{character()}}
}
\description{
Groovy Method Call
}
\seealso{
\url{http://docs.groovy-lang.org/docs/groovy-latest/html/documentation/#_named_parameters_2}
}
