% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groovy.R
\name{GroovyClosure}
\alias{GroovyClosure}
\title{Groovy Closure}
\usage{
GroovyClosure(..., parameters = character())
}
\arguments{
\item{...}{statements}

\item{parameters}{optional parameters: \code{character()}}
}
\description{
Groovy Closure
}
\seealso{
\url{https://groovy-lang.org/closures.html}
}
