% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groovy.R
\name{GString}
\alias{GString}
\title{Groovy String}
\usage{
GString(
  x,
  multiLine = grepl("\\n", x),
  interpolation = grepl("\\\\$\\\\{[^\\\\]*\\\\}", x)
)
}
\arguments{
\item{x}{\code{character} to use as the string base}

\item{multiLine}{allow string to span multiple lines}

\item{interpolation}{allow groovy string interpolation}
}
\description{
Groovy String
}
\seealso{
\url{https://groovy-lang.org/syntax.html#all-strings}
}
