% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/jenkins.R, R/job.R, R/view.R
\name{hasJob}
\alias{hasJob}
\alias{hasJob.Jenkins}
\alias{hasJob.JenkinsJob}
\alias{hasJob.JenkinsView}
\title{Check if a job exists}
\usage{
hasJob(x, ...)

\method{hasJob}{Jenkins}(x, name)

\method{hasJob}{JenkinsJob}(x, name)

\method{hasJob}{JenkinsView}(x, name)
}
\arguments{
\item{x}{object to check jobs for}
}
\value{
\code{logical}
}
\description{
S3 generic
}
