% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{hasView}
\alias{hasView}
\title{Check if a view exists}
\usage{
hasView(jenkins, name, ...)
}
\arguments{
\item{name}{name of the view to check for}

\item{...}{further arguments to \code{\link{jenkinsHEAD}}}
}
\description{
Check if a view exists
}
