% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indentLines}
\alias{indentLines}
\title{Indent lines}
\usage{
indentLines(text, indent = strrep(" ", getOption("rjenkins.indent", 4)))
}
\arguments{
\item{text}{\code{character()} vector of text snippets}

\item{indent}{\code{character()} string to indent with}
}
\value{
\code{character()}
}
\description{
Indent lines
}
