% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{install_jenkins}
\alias{install_jenkins}
\title{Install an exported artifact as package}
\usage{
install_jenkins(conn, job, pkg, build = "lastStableBuild",
  tmpDir = normalizePath(file.path("~", "Downloads")))
}
\arguments{
\item{conn}{result of \link{jenkinsConnection}}

\item{job}{job name}

\item{pkg}{name of the artifact that corresponds to the R package archive
to install}

\item{build}{build number of identifier}

\item{tmpDir}{temporary directory to store downloaded package archive}
}
\description{
Install an exported artifact as package
}
