% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_jenkins.R
\name{installPackageArtifact}
\alias{installPackageArtifact}
\title{Install a Package from Jenkins}
\usage{
installPackageArtifact(job, artifacts = listArtifacts(job,
  "lastSuccessfulBuild"), build = JENKINS_BUILD_REFS, latestOnly = TRUE,
  tmpDir = normalizePath(file.path("~", "Downloads")), ...)
}
\arguments{
\item{job}{object of class \code{jenkinsJob} e.g. created using \code{\link{getJob}}}

\item{artifacts}{\code{character()} vector or artifact name(s) that should be
installed as an R package. By default this is the complete list of artifacts
exported from the latest succesful build of the corresponding job.}

\item{build}{build number or a build ref. See \code{\link{JENKINS_BUILD_REFS}}}

\item{latestOnly}{only install the latest version when multiple versions
of the same package are given in \code{pkg}; \code{TRUE} by default}

\item{tmpDir}{temporary directory to store downloaded package archive}

\item{...}{further arguments to \code{\link{install.packages}}}
}
\value{
nothing
}
\description{
Install build artifacts exported from a jenkins job as R packages.
}
\details{
only artifacts that have a name that is interpretable under the
default package archive naming convention will be installed.
}
\seealso{
\code{\link{extractPackageArchives}} for filtering package names
from a list of build artifacts
}
