% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\docType{data}
\name{JENKINS_BUILD_REFS}
\alias{JENKINS_BUILD_REFS}
\title{Possible Jenkins Build Refs}
\format{An object of class \code{character} of length 6.}
\usage{
JENKINS_BUILD_REFS
}
\description{
a \code{character()} vector containing all possible build references
}
\details{
The following are valid build references: \itemize{
\item \code{lastBuild}
\item \code{lastCompletedBuild}
\item \code{lastFailedBuild}
\item \code{lastStableBuild}
\item \code{lastSuccessfulBuild}
\item \code{lastUnsuccessfulBuild}
}
}
\references{
https://wiki.jenkins.io/display/JENKINS/Terminology
}
\keyword{datasets}
