% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{jenkinsConnection}
\alias{jenkinsConnection}
\title{Connect to a Jenkins Server}
\usage{
jenkinsConnection(host, user, token)
}
\arguments{
\item{host}{jenkins endpoint}

\item{user}{user to connect with}

\item{token}{token to use for authentication}
}
\value{
object of class \code{jenkinsConnection}
}
\description{
Create a connection object for the Jenkins API.
}
