% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{jenkinsHEAD}
\alias{jenkinsHEAD}
\title{Execute a HEAD API request for Jenkins}
\usage{
jenkinsHEAD(jenkins, path, ...)
}
\arguments{
\item{jenkins}{See \code{\link{jenkins}}}

\item{path}{subpath to api}

\item{...}{other arguments to \code{\link[httr]{HEAD}}}
}
\description{
Execute a HEAD API request for Jenkins
}
