% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{jenkinsPOST}
\alias{jenkinsPOST}
\title{Execute a POST API request for Jenkins}
\usage{
jenkinsPOST(jenkins, path, xml = NULL, ...)
}
\arguments{
\item{jenkins}{See \code{\link{jenkins}}}

\item{path}{subpath}

\item{...}{other arguments to \code{\link[httr]{POST}}}
}
\description{
Execute a POST API request for Jenkins
}
