% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{listArtifacts}
\alias{listArtifacts}
\title{List Build Artifacts}
\usage{
listArtifacts(job, build = JENKINS_BUILD_REFS)
}
\arguments{
\item{job}{object of class \code{jenkinsJob} e.g. created using \code{\link{getJob}}}

\item{build}{build number or a build ref. See \code{\link{JENKINS_BUILD_REFS}}}
}
\description{
List Build Artifacts
}
