% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{listJobArtifacts}
\alias{listJobArtifacts}
\title{List artifacts}
\usage{
listJobArtifacts(conn, job, build = "lastStableBuild")
}
\arguments{
\item{conn}{result of \link{jenkinsConnection}}

\item{job}{job name}

\item{build}{build number of identifier}
}
\description{
List artifacts
}
