% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/generics.R, R/job.R
\name{listJobs.jenkinsConnection}
\alias{listJobs.jenkinsConnection}
\alias{listJobs}
\alias{listJobs.jenkinsJob}
\title{List job names}
\usage{
\method{listJobs}{jenkinsConnection}(x)

listJobs(x, ...)

\method{listJobs}{jenkinsJob}(x)
}
\arguments{
\item{x}{object to list jobs for}

\item{job}{object of class \code{jenkinsJob} e.g. created using \code{\link{getJob}}}
}
\description{
S3 generic
}
