% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{listViews}
\alias{listViews}
\title{List views}
\usage{
listViews(jenkins, ...)
}
\arguments{
\item{jenkins}{Jenkins server configuration. Result of \code{\link{jenkins}}}

\item{...}{further arguments to \code{\link{jenkinsGET}}}
}
\value{
vector of view names
}
\description{
List available views in the jenkins server
}
