% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\name{naryOp}
\alias{naryOp}
\title{N-ary operator}
\usage{
naryOp(n, parts)
}
\arguments{
\item{n}{exact number of arguments}

\item{parts}{vector of length \code{n +_1} to interlace with the arguments}
}
\value{
closure
}
\description{
N-ary operator
}
