% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobConfig.R
\name{pipelineConfig}
\alias{pipelineConfig}
\title{Create a new Pipeline}
\usage{
pipelineConfig(remote, credentialsId = NULL, pipeline = NULL)
}
\arguments{
\item{remote}{the git remote repository to check out}

\item{credentialsId}{credentials used to scan branches and check out sources}

\item{pipeline}{inline pipeline; see \code{\link{jenkinsPipeline}}}
}
\value{
XML document
}
\description{
Creates configuration for a pipeline job.
}
\details{
If an inline \code{pipeline} is supplied, the \code{remote}
argument will be ignored
}
\seealso{
\link{createMultiBranchPipeline} \link{createJob}
}
