% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\docType{data}
\name{pipelineDirectives}
\alias{pipelineDirectives}
\title{Pipeline Directives}
\format{
An object of class \code{list} of length 6.
}
\usage{
pipelineDirectives
}
\description{
Pipeline Directives
}
\references{
\url{https://www.jenkins.io/doc/book/pipeline/syntax/#declarative-directives}
}
\keyword{datasets}
