% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\docType{data}
\name{pipelineSections}
\alias{pipelineSections}
\title{Pipeline Sections}
\format{
An object of class \code{list} of length 4.
}
\usage{
pipelineSections
}
\description{
Sections in Declarative Pipeline typically contain one or more
\code{\link{pipelineDirectives}} or \code{\link{pipelineSteps}}.
}
\keyword{datasets}
