% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{print.Jenkins}
\alias{print.Jenkins}
\title{Show a Jenkins Object}
\usage{
\method{print}{Jenkins}(x, ...)
}
\arguments{
\item{x}{connection to a jenkins instance. Result of \code{\link{jenkins}}}

\item{...}{further arguments; not used}
}
\description{
Display information about a jenkins object.
}
