% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{print.jenkinsConnection}
\alias{print.jenkinsConnection}
\title{Show a Jenkins Object}
\usage{
\method{print}{jenkinsConnection}(x, ...)
}
\arguments{
\item{x}{connection to a jenkins instance. Result of \code{\link{jenkinsConnection}}}

\item{...}{further arguments; not used}
}
\description{
Display information about a jenkins object.
}
