% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{print.jenkinsJob}
\alias{print.jenkinsJob}
\title{Show a Jenkins Job}
\usage{
\method{print}{jenkinsJob}(x, ...)
}
\arguments{
\item{x}{object representing a jenkins job.}

\item{...}{further arguments; not used}
}
\description{
Show a Jenkins Job
}
\seealso{
\code{\link{getJob}}
}
