% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{print.JenkinsView}
\alias{print.JenkinsView}
\title{Show a Jenkins View}
\usage{
\method{print}{JenkinsView}(x, ...)
}
\arguments{
\item{x}{object representing a jenkins view.}

\item{...}{further arguments; not used}
}
\description{
Show a Jenkins View
}
