% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rightVariadicOp}
\alias{rightVariadicOp}
\title{Right-Variadic operator}
\usage{
rightVariadicOp(nf, nv, parts, left, sep, right)
}
\arguments{
\item{nf}{number of fixed arguments}

\item{nv}{minimum number of variable arguments}

\item{parts}{vector of length \code{n +_1} to interlace with the fixed arguments}

\item{left}{prefix for the variadic part}

\item{sep}{separator for the variadic part}

\item{right}{suffix for the variadic part}
}
\value{
closure
}
\description{
Right-Variadic operator
}
