% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{scheduleBuild}
\alias{scheduleBuild}
\title{Schedule a Build}
\usage{
scheduleBuild(job, params = NULL)
}
\arguments{
\item{job}{object of class \code{jenkinsJob} e.g. created using \code{\link{getJob}}}

\item{params}{build parameters; currently not supported yet. FIXME}
}
\description{
Schedule a Build
}
