% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{schedulePoll}
\alias{schedulePoll}
\title{Schedule SCM Polling}
\usage{
schedulePoll(job)
}
\arguments{
\item{job}{object of class \code{jenkinsJob} e.g. created using \code{\link{getJob}}}
}
\description{
TODO
}
