% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\name{stageDirective}
\alias{stageDirective}
\title{Stage Directive}
\usage{
stageDirective(stageName, ...)
}
\arguments{
\item{stageName}{stage name}
}
\value{
\code{character()}
}
\description{
DSL helper function
}
