% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\name{step}
\alias{step}
\title{Step}
\usage{
step(name, ...)
}
\arguments{
\item{name}{step name}
}
\value{
\code{character()}
}
\description{
DSL helper function
}
\examples{
\dontrun{
step("a", foo = 4, "5", bar = 77, foobar = FALSE)
}
}
