% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{summary.Jenkins}
\alias{summary.Jenkins}
\title{Print summary information about the Jenkins Server}
\usage{
\method{summary}{Jenkins}(object, ...)
}
\arguments{
\item{object}{connection to a jenkins instance. Result of \code{\link{jenkins}}}

\item{...}{further arguments; not used}
}
\description{
Print summary information about the Jenkins Server
}
