% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{summary.jenkinsConnection}
\alias{summary.jenkinsConnection}
\title{Print summary information about the Jenkins Server}
\usage{
\method{summary}{jenkinsConnection}(object, ...)
}
\arguments{
\item{object}{connection to a jenkins instance. Result of \code{\link{jenkinsConnection}}}

\item{...}{further arguments; not used}
}
\description{
Print summary information about the Jenkins Server
}
