% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{summary.jenkinsJob}
\alias{summary.jenkinsJob}
\title{Summarize Jenkins Job}
\usage{
\method{summary}{jenkinsJob}(job)
}
\arguments{
\item{job}{object of class \code{jenkinsJob} e.g. created using \code{\link{getJob}}}
}
\description{
Get summary information for the given jenkins job
}
