% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\name{variadicOp}
\alias{variadicOp}
\title{Variadic operator}
\usage{
variadicOp(n, left, sep, right)
}
\arguments{
\item{n}{minimum number of arguments}

\item{left}{text to use as a prefix}

\item{sep}{seperator text}

\item{right}{text to use as a suffix}
}
\value{
closure
}
\description{
Variadic operator
}
