% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJob.R
\name{createJob}
\alias{createJob}
\title{Create a new job}
\usage{
createJob(conn, name, config)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \code{\link{jenkinsConnection}}}

\item{name}{job name}

\item{config}{job xml specification given either as a file path or an object
of class \code{xml_document} from the \code{xml2} package}
}
\value{
jenkins job
}
\description{
Creates a new job with given name and xml config. Refer to
\link{createMultiBranchPipeline} and \link{createPipeline} for more
user-friendly wrappers.
}
\seealso{
\link[xml2]{read_xml} \link{crumbRequest}

\link{createMultiBranchPipeline} \link{createPipeline}
}
