% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJob.R
\name{createMultiBranchPipeline}
\alias{createMultiBranchPipeline}
\title{Create a new MultiBranch Pipeline}
\usage{
createMultiBranchPipeline(conn, name, remote, credentialsId)
}
\arguments{
\item{name}{multibranch pipeline name}

\item{remote}{the git remote repository to check out}

\item{credentialsId}{credentials used to scan branches and check out sources}
}
\value{
jenkins job
}
\description{
Creates a new multibranch pipeline with given name and
git remote details.
}
