% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{crumbRequest}
\alias{crumbRequest}
\title{Request a crumb}
\usage{
crumbRequest(conn)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \code{\link{jenkinsConnection}}}
}
\value{
crumb as \code{character()}
}
\description{
Request a crumb to deal with CSRF Protection
}
\references{
https://wiki.jenkins.io/display/JENKINS/Remote+access+API#RemoteaccessAPI-CSRFProtection
}
