% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groovy.R
\name{formatGroovyCall}
\alias{formatGroovyCall}
\title{Format Groovy Call}
\usage{
formatGroovyCall(call, brackets = TRUE, closureSugar = FALSE)
}
\arguments{
\item{call}{\code{\link{GroovyCall}}}

\item{brackets}{include brackets around arguments}

\item{closureSugar}{if the last argument is a closure, move it outside the
brackets}
}
\description{
Format Groovy Call
}
