% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/generics.R, R/job.R
\name{getJob.jenkinsConnection}
\alias{getJob.jenkinsConnection}
\alias{getJob}
\alias{getJob.jenkinsJob}
\title{Find a job}
\usage{
\method{getJob}{jenkinsConnection}(x, name)

getJob(x, name, ...)

\method{getJob}{jenkinsJob}(x, name)
}
\arguments{
\item{x}{job parent}

\item{name}{job name}
}
\value{
object of class \code{jenkinsJob}
}
\description{
S3 generic
}
