% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/jenkins.R, R/job.R, R/view.R
\name{getJob}
\alias{getJob}
\alias{getJob.Jenkins}
\alias{getJob.JenkinsJob}
\alias{getJob.JenkinsView}
\title{Find a job}
\usage{
getJob(x, name, ...)

\method{getJob}{Jenkins}(x, name, ...)

\method{getJob}{JenkinsJob}(x, name, ...)

\method{getJob}{JenkinsView}(x, name)
}
\arguments{
\item{x}{object to get job for}

\item{name}{job name}

\item{...}{zero or more child job names}
}
\description{
S3 generic
}
