% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/generics.R, R/job.R
\name{hasJob.jenkinsConnection}
\alias{hasJob.jenkinsConnection}
\alias{hasJob}
\alias{hasJob.jenkinsJob}
\title{Check if a job exists}
\usage{
\method{hasJob}{jenkinsConnection}(x, name)

hasJob(x, name)

\method{hasJob}{jenkinsJob}(x, name)
}
\arguments{
\item{x}{object to check jobs for}

\item{name}{job name}
}
\value{
\code{logical}
}
\description{
S3 generic
}
