% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{hasJob}
\alias{hasJob}
\title{Check if a job with given name exists}
\usage{
hasJob(conn, job)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \link{jenkinsConnection}}

\item{job}{job name}
}
\description{
Check if a job with given name exists
}
