% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{hasJob}
\alias{hasJob}
\title{Check if a job with given name exists}
\usage{
hasJob(conn, jobName)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \code{\link{jenkinsConnection}}}

\item{jobName}{job name}
}
\description{
Check if a job with given name exists
}
