% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\name{indentLines}
\alias{indentLines}
\title{Indent lines}
\usage{
indentLines(text, indent = "    ")
}
\arguments{
\item{text}{\code{character()} vector of text snippets ending in a newline}

\item{indent}{\code{character()} string to indent with}
}
\value{
\code{character()}
}
\description{
DSL helper function
}
