% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\name{indentLines}
\alias{indentLines}
\title{Indent lines}
\usage{
indentLines(text, indent = "    ")
}
\arguments{
\item{texts}{\code{character()} vector of text snippets ending in a newline}
}
\value{
\code{character()}
}
\description{
block header
}
