% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_jenkins.R
\name{install_jenkins}
\alias{install_jenkins}
\title{Install a Package from Jenkins}
\usage{
install_jenkins(uri, auth = jenkinsAuthEnv())
}
\arguments{
\item{uri}{resource identifier, formatted \code{<url>/<job>/<package>[#build]},
where build is optional and can be either a build number or a build ref. 
See \code{\link{JENKINS_BUILD_REFS}}. If no build is specified,
\code{lastSuccessfulBuild} will be used.}

\item{auth}{authentication credentials as returned by \link{jenkinsAuth}.
Can also be a named list, in which case the credentials under the name
which matches the hostname in \code{uri} will be used.}
}
\value{
package name (invisible)
}
\description{
Convenience wrapper to \code{\link{installPackageArtifacts}}
}
\examples{
\dontrun{
install_jenkins("https://ci.openanalytics.eu/packamon/packamon")
}
}
