% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{jenkinsConnection}
\alias{jenkinsConnection}
\title{Connect to a Jenkins Server}
\usage{
jenkinsConnection(url = "http://localhost", user = NULL,
  token = NULL, auth = NULL)
}
\arguments{
\item{url}{jenkins endpoint}

\item{user}{user to connect with}

\item{token}{token to use for authentication}

\item{auth}{authentication credentials obtained via \code{\link{jenkinsAuth}}}
}
\value{
object of class \code{jenkinsConnection}
}
\description{
Create a connection object for the Jenkins API. Credentials
must be specified either directly or through a configuration object.
}
