% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{jenkinsAuth}
\alias{jenkinsAuth}
\alias{jenkinsAuthEnv}
\alias{jenkinsAuthKeyringr}
\alias{jenkinsAuthFile}
\title{Jenkins Authentication Credentials}
\usage{
jenkinsAuth(user, token)

jenkinsAuthEnv(userVar = "JENKINS_USER", tokenVar = "JENKINS_TOKEN")

jenkinsAuthKeyringr(user, host = "localhost")

jenkinsAuthFile(credentialsFile = ".jenkins-credentials")
}
\arguments{
\item{user}{user to connect with}

\item{token}{token to use for authentication}

\item{userVar}{environment variable with user}

\item{tokenVar}{environment variable with token}

\item{host}{jenkins host identifier}

\item{credentialsFile}{connection or filename}
}
\value{
credentials; object of class \code{jenkinsCredentials}
}
\description{
\code{jenkinsAuth} provide credentials directly as
function arguments.

\code{jenkinsAuthEnv} source credentials from environment
variables

\code{jenkinsAuthKeyringr} use the \code{keyringr} package to lookup
an user password.

\code{jenkinsAuthFile} source credentials from file
}
