% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{jenkinsJob}
\alias{jenkinsJob}
\title{Get a jenkins job}
\usage{
jenkinsJob(conn, name)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \code{\link{jenkinsConnection}}}

\item{name}{job name}
}
\description{
Create an object representing a jenkins job.
}
