% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{listJobs}
\alias{listJobs}
\title{List all jobs names}
\usage{
listJobs(conn)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \link{jenkinsConnection}}
}
\description{
List all jobs names
}
