% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/jenkins.R, R/job.R, R/view.R
\name{listJobs}
\alias{listJobs}
\alias{listJobs.Jenkins}
\alias{listJobs.JenkinsJob}
\alias{listJobs.JenkinsView}
\title{List job names}
\usage{
listJobs(x, ...)

\method{listJobs}{Jenkins}(x, ...)

\method{listJobs}{JenkinsJob}(x)

\method{listJobs}{JenkinsView}(x, ...)
}
\arguments{
\item{x}{object to list jobs for}
}
\description{
S3 generic
}
