% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\name{jenkinsPipeline}
\alias{jenkinsPipeline}
\title{Generate Declarative Jenkins Pipeline syntax from a pipeline expression}
\usage{
jenkinsPipeline(pipelineExpr)
}
\arguments{
\item{pipelineExpr}{pipeline expression}
}
\value{
Declarative Pipeline syntax as a \code{character()}
}
\description{
TODO refer to Jenkins 2 book
}
