% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\name{jenkinsPipeline}
\alias{jenkinsPipeline}
\alias{pipeline}
\title{Declarative Jenkins Pipeline}
\usage{
jenkinsPipeline(pipelineExpr = pipeline())

pipeline(...)
}
\arguments{
\item{pipelineExpr}{pipeline expression}
}
\value{
Declarative Pipeline syntax as a \code{character()}
}
\description{
Generate Declarative Jenkins Pipeline syntax from a pipeline expression.
}
\examples{

if (interactive()) {
  
  x <- jenkinsPipeline(
      pipeline(
          agent("any"),
          stages(
              stage("Build", steps(echo("hello")))
          )
      )
  )
  
  # shorthand
  x <- pipeline(
      agent("any"),
      stages(
          stage("Build", steps(echo("hello")))
      )
  )
  
}
}
\references{
\url{https://jenkins.io/doc/book/pipeline/syntax/}
}
