% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\docType{data}
\name{pipelineDirectives}
\alias{pipelineDirectives}
\title{Pipeline directives}
\format{An object of class \code{list} of length 6.}
\usage{
pipelineDirectives
}
\description{
DSL helper functions
}
\keyword{datasets}
