% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\docType{data}
\name{pipelineSections}
\alias{pipelineSections}
\title{Pipeline sections}
\format{An object of class \code{list} of length 4.}
\usage{
pipelineSections
}
\description{
Pipeline sections
}
\keyword{datasets}
