% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineDSL.R
\docType{data}
\name{pipelineSteps}
\alias{pipelineSteps}
\title{Pipeline steps}
\format{
An object of class \code{list} of length 3.
}
\usage{
pipelineSteps
}
\description{
Pipeline steps
}
\references{
\url{https://jenkins.io/doc/pipeline/steps/}
}
\keyword{datasets}
