% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{print.jenkinsConnection}
\alias{print.jenkinsConnection}
\title{Show a Jenkins Object}
\usage{
\method{print}{jenkinsConnection}(conn)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \link{jenkins}}
}
\description{
Display information about a jenkins object.
}
