% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{print.JenkinsJob}
\alias{print.JenkinsJob}
\title{Show a Jenkins Job}
\usage{
\method{print}{JenkinsJob}(x, ...)
}
\arguments{
\item{x}{object representing a jenkins job.}

\item{...}{further arguments; not used}
}
\description{
Show a Jenkins Job
}
