% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkins.R
\name{summary.jenkinsConnection}
\alias{summary.jenkinsConnection}
\title{Print summary information about the Jenkins Server}
\usage{
\method{summary}{jenkinsConnection}(conn)
}
\arguments{
\item{conn}{connection to a jenkins instance. Result of \link{jenkins}}
}
\description{
Print summary information about the Jenkins Server
}
